Progress save: syscall parameter tested

+++ dev/kernel/inc/lib/list.h
@@ -6,7 +6,7 @@
 #include <err.h>
 #include <config.h>
 
-#define offsetof(TYPE,MEMBER)   ((size_t)&((TYPE*)0)->MEMBER)
+#define offsetof(TYPE,MEMBER)   ((uint)&((TYPE*)0)->MEMBER)
 
 struct List_s {
 	struct List_s* prev;
